Record runAt. Relates #7

+++ build/scaffold/stateScript.js
@@ -4,18 +4,20 @@
 
 const db = require("./connection");
 
-// this function shows state to the sequential-migrate cli
-const get = async () => {
-  // In this example, we check if the table exists because our first migration creates the table. If we migrate all the way down, and then check the status, the `get` function had better handle it
-  const doesTableExist = !!(
-    await db.query(
+// In this example, we check if the table exists because our first migration creates the table. If we migrate all the way down, the `get` and `delete` functions should understand what's going on.
+const doesTableExist = () =>
+  db
+    .query(
       `SELECT 1 as "column" FROM information_schema.tables WHERE table_name = 'migrations'`
     )
-  ).rows.length;
+    .then((res) => !!res.rows.length);
 
-  if (!doesTableExist) {
+// this function shows state to the sequential-migrate cli
+const get = async () => {
+  if (!(await doesTableExist())) {
     return [];
   }
+
   const stateRaw = await db.query("table public.migrations");
   return stateRaw.rows.map((state) => ({
     name: state.name,
@@ -37,6 +39,12 @@ const add = async (stateItem) => {
 
 // the sequential-migrate CLI calls this when it wants you to delete a migration from your store following a successful rollback
 const remove = async (stateItem) => {
+  if (!(await doesTableExist())) {
+    console.log(
+      `\nNot removing ${stateItem.name} from the store because the table does not exist...`
+    );
+    return;
+  }
   const query = "delete from public.migrations where name = $1";
   await db.query(query, [stateItem.name]);
 };
